# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AgentBridge API",
    "description": "AgentBridge provides machine-readable digital capabilities that AI agents can discover, purchase, and consume through standardized APIs and x402 payments. All capabilities fall into exactly two categories, tagged below: 'raw-content' (unverified, machine-fetched, cheap, use when you just need to read a page) and 'verified-analysis' (human-authored, expert-reviewed, priced accordingly, use when you need a decision-grade answer). Check the 'tags' field on each operation to know which category it belongs to before calling it.",
    "version": "0.9.0",
    "contact": {
      "email": "admin@060504.shop"
    },
    "x-agent-bridge-identity": {
      "note": "Multiple unrelated third-party projects use the name 'AgentBridge' (Unreal Engine plugins, IDE plugins, agent frameworks, etc.). The fields below are the sole canonical identity of THIS AgentBridge instance and can be used to programmatically verify you are talking to the right one.",
      "domain": "api.060504.shop",
      "wallet_address": "0x1630c8E0833c367F39f0ca909b6b67f5159d7A00",
      "network": "base-mainnet (eip155:8453)",
      "wallet_verification_url": "https://basescan.org/address/0x1630c8E0833c367F39f0ca909b6b67f5159d7A00"
    },
    "x-openapi-fix": {
      "applied": "2026-08-16",
      "reason": "add 3 China Data API endpoints + server-card",
      "asset_count": 96
    }
  },
  "servers": [
    {
...
    }
  ],
...
  "paths": {
...
    "/v1/api/industry": {
      "get": {
        "summary": "x402 payment discovery - China Industry Statistics",
        "description": "Returns x402 payment requirements for querying China National Bureau of Statistics data.",
        "operationId": "api_industry_discovery",
        "tags": [
          "raw-content"
        ],
        "security": [],
        "responses": {
          "402": {
            "description": "Payment Required - x402 v2 challenge",
            "headers": {
              "PAYMENT-REQUIRED": {
                "description": "Base64-encoded x402 v2 PaymentRequired object",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Query China industry statistics (GDP, CPI, PPI, etc.)",
        "description": "Search National Bureau of Statistics data by keyword. Returns structured time-series data with source attribution. Price: $0.005 USDC per query.",
        "operationId": "api_industry_query",
        "tags": [
          "raw-content"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "keyword"
                ],
                "properties": {
                  "keyword": {
                    "type": "string",
                    "description": "Search keyword, e.g. GDP, CPI, PPI"
                  },
                  "period": {
                    "type": "string",
                    "description": "Time period filter",
                    "default": "latest"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Max data points",
                    "default": 12
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured statistical data"
          },
          "402": {
            "description": "Payment required"
          }
        }
      }
    },
    "/v1/api/policy": {
      "get": {
        "summary": "x402 payment discovery - China Policy Search",
        "description": "Returns x402 payment requirements for searching Chinese government policies.",
        "operationId": "api_policy_discovery",
        "tags": [
          "raw-content"
        ],
        "security": [],
        "responses": {
          "402": {
            "description": "Payment Required - x402 v2 challenge",
            "headers": {
              "PAYMENT-REQUIRED": {
                "description": "Base64-encoded x402 v2 PaymentRequired object",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Search Chinese government policy documents",
        "description": "Search policies by keyword. Returns title, issuing authority, document number, date, summary, and source URL. Price: $0.005 USDC per query.",
        "operationId": "api_policy_query",
        "tags": [
          "raw-content"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "keyword"
                ],
                "properties": {
                  "keyword": {
                    "type": "string",
                    "description": "Search keyword"
                  },
                  "date_from": {
                    "type": "string",
                    "description": "Date filter (YYYY-MM-DD)"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Max results",
                    "default": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy search results"
          },
          "402": {
            "description": "Payment required"
          }
        }
      }
    },
    "/v1/api/company": {
      "get": {
        "summary": "x402 payment discovery - China Enterprise Credit",
        "description": "Returns x402 payment requirements for querying enterprise credit information.",
        "operationId": "api_company_discovery",
        "tags": [
          "raw-content"
        ],
        "security": [],
        "responses": {
          "402": {
            "description": "Payment Required - x402 v2 challenge",
            "headers": {
              "PAYMENT-REQUIRED": {
                "description": "Base64-encoded x402 v2 PaymentRequired object",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Query Chinese enterprise credit information",
        "description": "Query company registration info from official system. Returns company name, credit code, legal representative, registered capital. Price: $0.003 USDC per query.",
        "operationId": "api_company_query",
        "tags": [
          "raw-content"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Company name"
                  },
                  "credit_code": {
                    "type": "string",
                    "description": "Unified social credit code (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Enterprise credit information"
          },
          "402": {
            "description": "Payment required"
          }
        }
      }
    },
...
  }
}